rename vcsa to vcenter for clarity
VCSA stands for vCenter Server Appliance, and so is actually a vcenter.
vCenter is a much more popular terminology. Let's avoid any confusion
and just use just this therm.

In addition, we are about to introduce more VMware jobs, and the number
of nodes will be the distinguish factor. So when it makes sense, this
commit uses vcenter for clarity.