handle ansible-test with a role
Move playbooks/ansible-test-integration-base/run.yaml in two new roles:

- deploy-artifacts: fetch and install the artifacts (sdist and
  collections)
- ansible-test: run ansible-test

We also install Ara from the `ansible-test` role. To configure Ara,
we need to adjust the ansible.cfg file. The location of the file
is dynamic and the final location is only known in `ansible-test` role.
By delaying Ara configuration, we don't have to move the ansible.cfg
file to its final location anymore.

Signed-off-by: Gonéri Le Bouder <goneri@lebouder.net>