Add github-workflows job
All projects in github will have to run this job moving forward.
Basically, zuul will not be able to merge PRs with contain changes to
these files (.github/workflows). This appears to be a security feature
in github, but break automated gating in of commits in zuul.  This is
still true, if we completly disable github actions on repos.

Signed-off-by: Paul Belanger <pabelanger@redhat.com>