[DNM][WIP] Add vqfx nodesets and jobs to Ansible Zuul
This commit adds several vqfx nosetes with the different python
combinations to ansible zuul jobs. It also creates the different jobs
for testing with netconf.

Note: this patch will be depending on junos_filters/acls resource
module, which would need vqfx in order for it to be tested. Once that is
done, this commit zuul templates will be modified in order for it to:

1) Get added to check/gate queues
2) Be triggered only on junos_filters/acls files changes

Depends-On: https://github.com/ansible-network/network-runner/pull/38
Signed-off-by: Daniel Mellado <dmellado@redhat.com>