nodesets: add appliance_no_ssh group
`add-build-sshkey` from `base-minimal` will regenerate a new set
of SSH key and push them on the nodes. However it ignores the
hosts from the `appliance` group. As a result, there is no way to
access the appliances over SSH after this step.

The `appliance` group is too broad. This new group resolves the
problem. It clearly identifies the hosts that should not be targeted
by this operation.

We currently face the problem with the vcenter and splunk hosts. So
far the workaround was to hardcode a known password to the zuul users.