vmware: write /etc/hosts after hostname change
The `esxcli` command resolve the system hostname to know the address of
the local server. If the resolution fails, the typical error if:

```
IO error: [Errno 97] Address family not supported by protocol
```

If we change `/etc/hosts` BEFORE we fix the system hostname, we can
break the name resolution.

See: https://github.com/ansible-collections/vmware/issues/144