network: move py{35,37,38} jobs in perodical pipeline
There is currently quite a few running jobs but they consume wy too much jobs
and as a result, we are often at capacity. Our resources are not unlimited,
and in this case, jobs just stay pending... for hours.

For instance, we've got currently a vmware job that has been running
for 16h now.

This kind of delay does not benefit anyone, they impact our moral,
the contributors motivations and they certainly don't improve the quality
of the collections.

Python35, Python37 and Python38 are not criticial and consume a lot of
resources. They can definitiely be run in a delayed pipeline. And this would
help use greatly to speed up our CI.