vSphere7: ensure TCP timestamps are ON
If the ESXi and the vSphere are on the same hypervisor, the network
becomes really fast and we can be hit by unpredicable packet order.
The timestamps is used by PAWS (TCP Sequence number wrapping) to
detect this situation. This is the reason why we want to keep it on.

See: https://tejparkash.wordpress.com/2010/12/05/paws-tcp-sequence-number-wrapping-explained/