aws: introduce ansible-test-splitter job
This commit changes the way we split up the AWS jobs. We've got now a
first job called `ansible-test-splitter` that is in charge of:

- identify what target needs to be run (like the `ansible-test` role)
- split up the workload on different jobs

This logic is handled by a new role called `ansible-test-splitter` and
these two tasks are done through two Python scripts.

Why Python? Well, the two tasks were getting overcomplicated and the
use of Python simplify a lot the troubleshoot and the code complexity.
In addition, we can later create action plugin from these scripts.

Finally, the commit also introduces a semaphore for the AWS jobs. It's
a way to control the number of paralel jobs.