Ensure splitter script runs all requested targets (#1091)
Ensure splitter script runs all requested targets

This fixes a couple problems with the splitter script. The first problem
is that modules listed in aliases files were not being run. The second
problem is that non-slow targets may be silently skipped in cases when
there are six or more slow jobs being run.

Reviewed-by: Gonéri Le Bouder <goneri@lebouder.net>
Reviewed-by: Mike Graves <mgraves@redhat.com>
Reviewed-by: None <None>