Add amazon.cloud project (#1403)
Add amazon.cloud project

This adds jobs for the new amazon.cloud project. We decided we are only going to do integration testing for a handful of modules, so there is no need to use the splitter.
I'm not really sure I know what I'm doing here, so this may be the wrong way to do things.

Reviewed-by: Gonéri Le Bouder <goneri@lebouder.net>