community.vmware: Enable fail-fast (#1757)
community.vmware: Enable fail-fast

The integration tests for community.vmware sometimes fail for no obvious reason. That is, not because there's something wrong with the code but because of a network issue or something. But even if one of the jobs fails, the others still keep running.
This means that even if a voting test job fails, another one might run for hours because of retries. But the CI still fails as a whole.
I think the CI pipeline should fail immediately in this case. This would both give us a quicker response and save resources.