Run community.vmware integration tests with 2.14 (#1766)
Run community.vmware integration tests with 2.14

It looks like community.vmware integration tests are still run with ansible (core) 2.12. I think it's about time we change this to the current version 2.14.
I hope I've implemented this correctly, I'm not an expert on ansible-zuul-jobs.
edit: Especially since the sanity tests are run with 2.13 and 2.14, but not with 2.12.