community.vmware: Make ansible-galaxy-importer non-voting (#1901)
It looks like ansible-galaxy-importer has problems with doc fragments from another collection. So I think we should make it non-voting and test this with a GHA.
At least until it's fixed or I find a way to make it work.
FYI I ran into this when trying to re-use doc fragments from vmware.vmware in community.vmware: ansible-collections/community.vmware#2369

Reviewed-by: Alina Buzachis
Reviewed-by: Mike Graves <mgraves@redhat.com>